pyproject.toml
[tool.poetry]
name = "zonetool"
version = "0.1.0"
description = ""
authors = ["j-minster <jack@minster.com.au>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
geopandas = "^0.14.3"
click = "^8.1.7"
dbf = "^0.99.9"

[tool.poetry.scripts]
main = "zonetool.run.main:run"


[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.1.2"
icecream = "^2.1.3"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
